fixing bugs and updating serializer
