+++ src/researchhub_comment/serializers/constants/rh_comment_serializer_contants.py
@@ -1,5 +1,6 @@
 RH_COMMENT_FIELDS = [
     "children",
+    "children_count",
     "comment_content_json",
     "comment_content_type",
     "context_title",
@@ -7,6 +8,7 @@ RH_COMMENT_FIELDS = [
     "created_by",
     "updated_by",
     "id",
+    "is_accepted_answer",
     "is_edited",
     "parent",
     "thread",
@@ -15,6 +17,7 @@ RH_COMMENT_FIELDS = [
 RH_COMMENT_READ_ONLY_FIELDS = [
     "children",
     "created_date",
+    "is_accepted_answer",
     "id",
     "is_edited",
     "updated_date",
